Fix a data loss window in Journal#roll().
Since Journal#roll() renames the existing journal and writes a new one
to the original path, if the JVM or system goes down while the new
journal is being written then it may not have all the entries that the
original journal does.  So write the new journal into a temporary path
and then rename it into the journal location.